requirements.txt
pyspark==3.5.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
apache-airflow==2.8.1
